Added Input and Output Signal Queues on client side. Changed variable types from int to byte StatusType.
git-svn-id: file:///e/SVN/serious_messenger@21 b1debafe-a605-4e1b-80f9-83a58a4e6cd1

+++ Serious Client/src/client/StatusType.java
@@ -2,12 +2,12 @@ package client;
 
 public class StatusType {
 	
-	final public static int Offline = 0;
-	final public static int Online = 1;
-	final public static int Idle = 2;
-	final public static int Away = 3;
-	final public static int Busy = 4;
-	final public static int BeRightBack = 5;
+	final public static byte Offline = 0;
+	final public static byte Online = 1;
+	final public static byte Idle = 2;
+	final public static byte Away = 3;
+	final public static byte Busy = 4;
+	final public static byte BeRightBack = 5;
 	
 	public static boolean isValid(int status) {
 		return status >= 0 && status <= 5;
